combine flip_social_alloc and alloc
