+++ alloc/alloc.h
@@ -2,6 +2,7 @@
 #include <flip_social.h>
 #include <callback/callback.h>
 #include <free/free.h>
+FlipSocialApp *flip_social_app_alloc();
 void auth_headers_alloc(void);
 FlipSocialFeedMini *flip_feed_info_alloc(void);
 bool messages_dialog_alloc(bool free_first);
